# node-output/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ScreenSpace",
        "description": "This is a non commercial project for a software engineering degree. It´s a REST API for cinema management and movie ticket booking mobile app.",
        "contact": {
            "email": "contact@screenspace.com"
        },

        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://virtserver.swaggerhub.com/BREUERANDRES/ScreenSpace-API/1.0.0",
            "description": "SwaggerHub API Auto Mocking"
        },
        {
            "url": "http://localhost:3500/api/v1",
            "description": "Developement URL"
        },
        {
            "url": "https://screenspace-api.azurewebsites.net/api/v1",
            "description": "Production URL"
        }
    ],
...
    "paths": {
...
        "/users/": {
            "get": {
                "tags": ["Users"],
                "summary": "Get logged user data",
                "description": "By passing in the appropriate options, you can search for available specific user",
                "operationId": "searchUserById",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_2"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "bad input parameter"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": ["Users"],
                "summary": "Delete logged user",
                "description": "This can only be done by the logged in user.",
                "operationId": "deleteUser",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": ["Users"],
                "summary": "Update logged user",
                "description": "This can only be done by the logged in user.",
                "operationId": "updateUser",
                "requestBody": {
                    "description": "Update an existent user in the store",
                    "content": {
                        "x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/users_body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_2"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/owners": {
            "get": {
                "tags": ["Owners"],
                "summary": "Get logged owners data",
                "description": "Gets owners personal information. Must be logged in.",
                "operationId": "getOwner",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_5"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Owners"],
                "summary": "Sign Up",
                "description": "Create a new account for owner and send email for validation",
                "operationId": "addOwner",
                "requestBody": {
                    "description": "Signup only needs email and password",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/owners_body"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_6"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Input Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_400_2"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": ["Owners"],
                "summary": "Delete logged Owner",
                "description": "Owner is modified to inactive.",
                "operationId": "deleteOwner",
                "responses": {
                    "204": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_204"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": ["Owners"],
                "summary": "Update logged Owner",
                "description": "This can only be done by the logged in owner.",
                "operationId": "updateOwner",
                "requestBody": {
                    "description": "Update an existent Owner. Can update one or more properties",
                    "content": {
                        "x-www-form-urlencoded": {
                            "schema": {
                                "$ref": "#/components/schemas/owners_body_1"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_5"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/cinemas": {
            "get": {
                "tags": ["Cinemas"],
                "summary": "Searches all cinemas",
                "description": "Searches all cinemas",
                "operationId": "searchCinema",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_9"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Cinemas"],
                "summary": "Create new cinema",
                "description": "Create new cinema. Must be done by a logged owner",
                "operationId": "addCinema",
                "requestBody": {
                    "description": "Create a new cinema",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/cinemas_body"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_201"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Cinema alredy exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_400_4"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/theaters": {
            "post": {
                "tags": ["Theaters"],
                "summary": "Create new Theater",
                "description": "Create new Theater",
                "operationId": "addTheater",
                "requestBody": {
                    "description": "Create a new Theater",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/theaters_body"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_201_1"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Theater alredy exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_400_4"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/showtimes": {
            "get": {
                "tags": ["Showtimes"],
                "summary": "Searches Showtimes with query params",
                "description": "Gets all showtimes with query options",
                "operationId": "searchShowtime",
                "parameters": [
                    {
                        "name": "startTime",
                        "in": "query",
                        "description": "Filter by date",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available"
                        }
                    },
                    {
                        "name": "movie",
                        "in": "query",
                        "description": "Filter by movie",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available"
                        }
                    },
                    {
                        "name": "cinema",
                        "in": "query",
                        "description": "Filter by cinema",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "default": "available"
                        }
                    },
                    {
                        "name": "availableSeats",
                        "in": "query",
                        "description": "Filter by available seats left",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_11"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403_2"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Showtimes"],
                "summary": "Create new Showtime",
                "description": "Create new Showtime",
                "operationId": "addShowtimes",
                "requestBody": {
                    "description": "Create a new Showtime. Only available for logged owners",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/showtimes_body"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_201_2"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Showtime alredy exists",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_400_4"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/movies": {
            "get": {
                "tags": ["movies"],
                "summary": "Searches all movies",
                "description": "Search movies with query parameters",
                "operationId": "searchmovies",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit the size of movies",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "example": "10"
                        }
                    },
                    {
                        "name": "skip",
                        "in": "query",
                        "description": "For pagination",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "example": "2"
                        }
                    },
                    {
                        "name": "genre",
                        "in": "query",
                        "description": "Filter by genre",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string",
                            "example": "Terror"
                        }
                    },
                    {
                        "name": "premiered",
                        "in": "query",
                        "description": "Filter by premiered",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "cinema",
                        "in": "query",
                        "description": "Filter by cinema",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "rating",
                        "in": "query",
                        "description": "Order by rating",
                        "required": false,
                        "style": "form",
                        "explode": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_13"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403_2"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/bookings": {
            "get": {
                "tags": ["Bookings"],
                "summary": "Searches user bookings",
                "description": "Get user bookings. You must be logged in",
                "operationId": "searchBooking",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_200_14"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403_2"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": ["Bookings"],
                "summary": "New showtime booking",
                "description": "Create new Booking",
                "operationId": "addBooking",
                "requestBody": {
                    "description": "Create a new Booking",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/bookings_body"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_201_3"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/ratings": {
            "post": {
                "tags": ["Ratings"],
                "summary": "Rate a movie",
                "description": "Rate a movie",
                "operationId": "rateMovie",
                "requestBody": {
                    "description": "Create a new rating",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ratings_body"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_201_4"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Invalid Credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/inline_response_403"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "inline_response_403": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "You must be logged in"
                    }
                }
            },
            "inline_response_200_2": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "$ref": "#/components/schemas/inline_response_200_2_data"
                    }
                }
            },
            "users_body": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Tom Jones"
                    },
                    "imageUrl": {
                        "type": "string",
                        "format": "url",
                        "example": "https://....."
                    }
                }
            },
...
            "inline_response_200_5": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "data": {
                        "$ref": "#/components/schemas/inline_response_200_5_data"
                    }
                }
            },
            "owners_body": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "example": "user@mail.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "pass1234"
                    }
                }
            },
            "inline_response_200_6": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    },
                    "message": {
                        "type": "string",
                        "example": "Confirmation email sent!"
                    }
                }
            },
            "inline_response_400_2": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "fail"
                    },
                    "message": {
                        "type": "string",
                        "example": "Enter a valid email"
                    }
                }
            },
            "inline_response_204": {
                "type": "object",
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "success"
                    }
                }
            },
            "owners_body_1": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Tom"
                    },
                    "mail": {
                        "type": "string",
                        "format": "mail",
                        "example": "tomjones@gmail.com"
                    },
                    "imageUrl": {
                        "type": "string",
                        "format": "url",
                        "example": "https://....."
                    }
                }
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}